requirements.txt
opencv-python>=4.5.0,<5.0.0
pandas>=1.3.0,<3.0.0
numpy>=1.21.0,<2.0.0
torch>=1.9.0,<2.0.0
timm>=0.6.0,<1.0.0
albumentations>=1.1.0,<2.0.0
pyyaml>=6.0,<7.0.0
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
python-multipart>=0.0.6
pillow>=9.0.0,<11.0.0
tqdm>=4.65.0,<5.0.0
openpyxl>=3.1.0,<4.0.0
slowapi>=0.1.9,<1.0.0
